refactor: 팀 패키지 패키지 컨벤션에 맞추도록 변경
