+++ src/main/java/com/example/sidemanagementbe/web/security/config/OAuth2ClientRegistrationRepositoryConfiguration.java
@@ -1,4 +1,4 @@
-package com.example.sidemanagementbe.login.config;
+package com.example.sidemanagementbe.web.security.config;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -16,8 +16,7 @@ import org.springframework.security.oauth2.client.registration.ClientRegistratio
 import org.springframework.security.oauth2.client.registration.InMemoryClientRegistrationRepository;
 
 /**
- * OAuth2 클라이언트의 등록 정보를 관리하기 위한 설정 클래스 Spring Security의 OAuth2 클라이언트를 사용하여 외부 인증 서비스(예: Google, Facebook, Kakao 등)와 연동할
- * 때 사용됨
+ * OAuth2 클라이언트의 등록 정보를 관리하기 위한 설정 클래스 Spring Security의 OAuth2 클라이언트를 사용하여 외부 인증 서비스(예: Google, Facebook, Kakao 등)와 연동할 때 사용됨
  */
 @Configuration
 @EnableConfigurationProperties(OAuth2ClientProperties.class)
@@ -33,8 +32,7 @@ public class OAuth2ClientRegistrationRepositoryConfiguration {
 
     /**
      * OAuth2 클라이언트 등록 정보를 관리하는 ClientRegistrationRepository 인터페이스의 구현체인 InMemoryClientRegistrationRepository를 생성.
-     * InMemoryClientRegistrationRepository는 메모리에 클라이언트 등록 정보를 저장하고 필요할 때 해당 정보를 제공. 이를 통해 애플리케이션에서 OAuth2 클라이언트 등록 정보에
-     * 쉽게 접근할 수 있음.
+     * InMemoryClientRegistrationRepository는 메모리에 클라이언트 등록 정보를 저장하고 필요할 때 해당 정보를 제공. 이를 통해 애플리케이션에서 OAuth2 클라이언트 등록 정보에 쉽게 접근할 수 있음.
      */
     @Bean
     @ConditionalOnMissingBean(ClientRegistrationRepository.class)
